Started retrieve images

+++ app/src/main/java/com/zukkadev/it/flickrtourist/model/FlickrImages.java
@@ -12,29 +12,29 @@ public class FlickrImages {
     private int id;
 
     @ColumnInfo(name = "image_id")
-    private int ImageID;
+    private long ImageID;
 
     @ColumnInfo(name = "image_url")
     private String ImageURL;
 
-    public FlickrImages(int id, int ImageID, String ImageURL) {
+    public FlickrImages(int id, long ImageID, String ImageURL) {
         this.id = id;
         this.ImageID = ImageID;
         this.ImageURL = ImageURL;
     }
 
     @Ignore
-    public FlickrImages(int ImageID, String ImageURL) {
+    public FlickrImages(long ImageID, String ImageURL) {
         this.ImageID = ImageID;
         this.ImageURL = ImageURL;
     }
 
     public int getId() { return id; }
-    public int getImageID() { return ImageID; }
+    public long getImageID() { return ImageID; }
     public String getImageURL() { return ImageURL; }
 
     public void setId(int id) { this.id = id; }
-    public void setImageID(int imageID) { this.ImageID = imageID; }
+    public void setImageID(long imageID) { this.ImageID = imageID; }
     public void setImageURL(String imageURL) { this.ImageURL = imageURL; }
 
 }
